[#29] Feat : 코드 리팩토링
- 에러메세지 출력 및 성공 메세지 출력

- 생성, 수정, 삭제, 이동할 때, 변경된 데이터를 전달

- title값이 전달되지 않더라도 default값으로 New Card가 자동 생성되게
  수정

-------------------------
-------------------------

+++ BE/src/main/java/todo3/codesquad/Controller/TodoController.java
@@ -8,11 +8,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import todo3.codesquad.domain.Response;
+import todo3.codesquad.message.FailedMessage;
 import todo3.codesquad.message.ResponseMessage;
 import todo3.codesquad.domain.Card;
 import todo3.codesquad.domain.CardRepository;
 import todo3.codesquad.domain.Col;
 import todo3.codesquad.domain.ColRepository;
+import todo3.codesquad.message.SuccessMessage;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -28,25 +30,33 @@ public class TodoController {
     CardRepository cardRepository;
 
     @PostMapping("/api/cards")
-    public void createCard(@RequestBody Map<String, Object> map) {
+    public ResponseEntity<ResponseMessage> createCard(@RequestBody Map<String, Object> map) {
         Col col = colRepository.findByColName(map.get("colName").toString()).orElse(null);
+        if(col == null){
+            return new ResponseEntity<>(new ResponseMessage(FailedMessage.NULL_DATA_MESSAGE,null), HttpStatus.NOT_FOUND);
+        }
         List<Card> cards = col.getCards();
         Card card = new Card(map);
         cards.add(card);
         colRepository.save(col);
+        return new ResponseEntity<>(new ResponseMessage(SuccessMessage.SUCCESS_CREATE,card), HttpStatus.OK);
     }
 
     @PostMapping("/api/cards/update")
-    public void updateCard(@RequestBody Map<String, Object> map) {
+    public ResponseEntity<ResponseMessage> updateCard(@RequestBody Map<String, Object> map) {
         Col col = colRepository.findByColName(map.get("colName").toString()).orElse(null);
+        if(col == null){
+            return new ResponseEntity<>(new ResponseMessage(FailedMessage.NULL_DATA_MESSAGE,null), HttpStatus.NOT_FOUND);
+        }
         List<Card> cards = col.getCards();
         Card card = cards.get((int) map.get("row") - 1);
         card.update(map);
         colRepository.save(col);
+        return new ResponseEntity<>(new ResponseMessage(SuccessMessage.SUCCESS_UPDATE,card), HttpStatus.OK);
     }
 
     @PostMapping("/api/cards/move")
-    public void moveCard(@RequestBody Map<String, Object> map) {
+    public ResponseEntity<ResponseMessage> moveCard(@RequestBody Map<String, Object> map) {
         String originColName = map.get("originColName").toString();
         String destinationColName = map.get("destinationColName").toString();
         int originRow = Integer.parseInt(map.get("originRow").toString());
@@ -54,15 +64,17 @@ public class TodoController {
 
         Col originCol = colRepository.findByColName(originColName).orElse(null);
         Col destinationCol = colRepository.findByColName(destinationColName).orElse(null);
-
+        if(originCol == null || destinationCol == null){
+            return new ResponseEntity<>(new ResponseMessage(FailedMessage.NULL_DATA_MESSAGE,null), HttpStatus.NOT_FOUND);
+        }
         List<Card> originCards = originCol.getCards();
         List<Card> destinationCards = destinationCol.getCards();
 
         if (originRow > originCards.size()) {
-            System.out.println("멀 옴기는거야?");
+            return new ResponseEntity<>(new ResponseMessage(FailedMessage.SIZE_ERROR_MESSAGE,null), HttpStatus.BAD_REQUEST);
         }
         if (destinationRow > destinationCards.size()) {
-            System.out.println("어디다 옴기는거야?");
+            return new ResponseEntity<>(new ResponseMessage(FailedMessage.SIZE_ERROR_MESSAGE,null), HttpStatus.BAD_REQUEST);
         }
         Card movedCard = originCards.get(originRow - 1);
         originCards.remove(originRow - 1);
@@ -79,13 +91,18 @@ public class TodoController {
 
         colRepository.save(originCol);
         colRepository.save(destinationCol);
+        return new ResponseEntity<>(new ResponseMessage(SuccessMessage.SUCCESS_MOVE,movedCard), HttpStatus.OK);
     }
 
     @PostMapping("/api/cards/delete")
-    public void deleteCard(@RequestBody Map<String, Object> map) {
+    public ResponseEntity<ResponseMessage> deleteCard(@RequestBody Map<String, Object> map) {
         Card card = cardRepository.findById(Long.parseLong(map.get("id").toString())).orElse(null);
+        if(card == null){
+            return new ResponseEntity<>(new ResponseMessage(FailedMessage.NULL_DATA_MESSAGE,null), HttpStatus.NOT_FOUND);
+        }
         card.delete();
         cardRepository.save(card);
+        return new ResponseEntity<>(new ResponseMessage(SuccessMessage.SUCCESS_DELETE,card), HttpStatus.OK);
     }
 
     @GetMapping("/api/cards/show")
@@ -99,8 +116,6 @@ public class TodoController {
             Response response = new Response(columnName,tempCards);
             resultResponse.add(response);
         }
-
-
-        return new ResponseEntity<>(new ResponseMessage(resultResponse), HttpStatus.OK);
+        return new ResponseEntity<>(new ResponseMessage(SuccessMessage.SUCCESS_SHOW,resultResponse), HttpStatus.OK);
     }
 }
